Rename files to keep consistency
