+++ cpp/velox/compute/VeloxBackend.h
@@ -27,11 +27,11 @@
 
 #include "VeloxToRowConverter.h"
 #include "arrow/c/abi.h"
-#include "compute/exec_backend.h"
+#include "compute/Backend.h"
 #include "include/arrow/c/bridge.h"
-#include "memory/columnar_batch.h"
-#include "memory/velox_columnar_batch.h"
-#include "memory/velox_memory_pool.h"
+#include "memory/ColumnarBatch.h"
+#include "memory/VeloxColumnarBatch.h"
+#include "memory/VeloxMemoryPool.h"
 #include "substrait/algebra.pb.h"
 #include "substrait/capabilities.pb.h"
 #include "substrait/extensions/extensions.pb.h"
@@ -40,7 +40,7 @@
 #include "substrait/plan.pb.h"
 #include "substrait/type.pb.h"
 #include "substrait/type_expressions.pb.h"
-#include "utils/metrics.h"
+#include "utils/Metrics.h"
 #include "velox/common/file/FileSystems.h"
 #include "velox/connectors/hive/FileHandle.h"
 #include "velox/connectors/hive/HiveConnector.h"
