+++ cpp/velox/compute/VeloxBridge.cc
@@ -1,5 +1,5 @@
 
-#include "bridge.h"
+#include "VeloxBridge.h"
 
 #include <algorithm>
 #include <cerrno>
