+++ cpp/velox/compute/VeloxToRowConverter.h
@@ -21,8 +21,8 @@
 #include <arrow/record_batch.h>
 #include <arrow/type.h>
 
-#include "operators/c2r/arrow_columnar_to_row_converter.h"
-#include "operators/c2r/columnar_to_row_base.h"
+#include "operators/c2r/ArrowColumnarToRowConverter.h"
+#include "operators/c2r/ColumnarToRow.h"
 #include "velox/vector/ComplexVector.h"
 
 namespace gluten {
